pyproject.toml
[tool.poetry]
name = "posting"
version = "0.1.0"
description = ""
authors = ["Mateus Diniz <serv.mateus@gmail.com>"]
readme = "README.md"
packages = [
    { include = "./core/posting" }
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
factory-boy = "^3.3.0"
djangorestframework = "^3.14.0"
pytest-django = "^4.8.0"
flake8 = "^7.0.0"
black = "^24.2.0"
django-extensions = "^3.2.3"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
